refactor: unify client/server type hierarchy
- Create single File, User, Dependencie types at src/lib/types/
- Remove duplicate client/ and server/ type directories
- Update all ~30 import paths across the project
- Remove Schema generics from Mongoose models
- Improve typing: createdAt/updatedAt as string instead of boolean | string
