+++ src/lib/state/index.ts
@@ -1,4 +1,4 @@
-import { File } from 'src/lib/types/client/File'
+import { File } from 'src/lib/types/File'
 
 export const fileInitialState: File = {
 	json: {
