+++ src/lib/types/Dependencie.ts
@@ -1,0 +1,6 @@
+export type Dependencie = {
+	id?: string
+	name: string
+	version?: string
+	description?: string
+}
