+++ src/lib/types/User.ts
@@ -1,0 +1,8 @@
+export interface User {
+	id: string
+	avatar?: string
+	username?: string
+	email?: string
+	createdAt?: string
+	updatedAt?: string
+}
