--- src/lib/types/client/Dependencie.d.ts
@@ -1,6 +1,0 @@
-export type Dependencie = {
-    id?: string, 
-    name: string, 
-    version?: string, 
-    description?: string
-}
