+++ src/lib/types/client/hooks/UseDependenciesProps.d.ts
@@ -1,4 +1,4 @@
-import { File } from 'src/lib/types/client/File'
+import { File } from 'src/lib/types/File'
 
 export type UseDependenciesProps = {
     classType: string,
