+++ src/lib/types/client/hooks/UseScriptsProps.d.ts
@@ -1,4 +1,4 @@
-import { File } from 'src/lib/types/client/File'
+import { File } from 'src/lib/types/File'
 
 export type UseScriptsProps = {
