--- src/lib/types/server/User.d.ts
@@ -1,8 +1,0 @@
-export interface User {
-    _id: string,
-    avatar?: string, 
-	username?: string, 
-	email?: string,
-    createdAt: boolean | string,
-    updatedAt: boolean | string
-}
