+++ src/pages/api/dependencies/[id].ts
@@ -1,4 +1,4 @@
-import { Dependencie } from 'src/lib/types/client/Dependencie'
+import { Dependencie } from 'src/lib/types/Dependencie'
 import axios from 'axios'
 import type { NextApiRequest, NextApiResponse } from 'next'
