+++ src/pages/api/users/[id]/index.ts
@@ -1,4 +1,4 @@
-import { User } from 'src/lib/types/server/User'
+import { User } from 'src/lib/types/User'
 import UserNotExistError from 'src/lib/errors/server/UserNotExist'
 import UserRepository from 'src/lib/database/repository/UserRepository'
 import type { NextApiRequest, NextApiResponse } from 'next'
