+++ src/reducer/index.ts
@@ -1,4 +1,4 @@
-import { File } from 'src/lib/types/client/File'
+import { File } from 'src/lib/types/File'
 import { FileAction } from 'src/lib/types/client/reducer/FileAction'
 
 const fileInitialState: object = {
