+++ src/services/deleteUserFile.ts
@@ -1,5 +1,5 @@
 import { DELETE_USER_FILE_ENDPOINT } from 'src/lib/util/constants'
-import { File } from 'src/lib/types/client/File'
+import { File } from 'src/lib/types/File'
 import callApi from 'src/lib/util/callApi'
 
 export default async (userId: string, fileId: string): Promise<File> => {
