+++ src/services/getUser.ts
@@ -1,5 +1,5 @@
 import { GET_USER_ENDPOINT } from 'src/lib/util/constants'
-import { User } from 'src/lib/types/client/User'
+import { User } from 'src/lib/types/User'
 import callApi from 'src/lib/util/callApi'
 
 export default async (userId: string): Promise<User> => {
