+++ src/services/updateFile.ts
@@ -1,4 +1,4 @@
-import { File } from 'src/lib/types/client/File'
+import { File } from 'src/lib/types/File'
 import { UPDATE_FILE_ENDPOINT } from 'src/lib/util/constants'
 import callApi from 'src/lib/util/callApi'
